+++ src/cglsr.jl
@@ -74,7 +74,7 @@ res = predict(fm, Xtest) ;
 res.pred
 rmsep(ytest, res.pred)
 f, ax = scatter(vec(res.pred), ytest)
-abline!(ax, 0, 1)
+ablines!(ax, 0, 1)
 f
 ```
 """