+++ src/cplsr_avg.jl
@@ -57,7 +57,7 @@ res = predict(fm, Xtest)
 res.posterior
 rmsep(res.pred, ytest)
 f, ax = scatter(vec(res.pred), ytest)
-abline!(ax, 0, 1)
+ablines!(ax, 0, 1)
 f
 ```
 """