+++ src/knnr.jl
@@ -63,7 +63,7 @@ fm = knnr(Xtrain, ytrain; nlvdis = nlvdis,
 res = predict(fm, Xtest)
 rmsep(res.pred, ytest)
 f, ax = scatter(vec(pred), ytest)
-abline!(ax, 0, 1)
+ablines!(ax, 0, 1)
 f
 ```
 """