+++ src/lwplsr.jl
@@ -102,7 +102,7 @@ fm = lwplsr(Xtrain, ytrain; nlvdis = nlvdis,
 res = predict(fm, Xtest)
 rmsep(res.pred, ytest)
 f, ax = scatter(vec(res.pred), ytest)
-abline!(ax, 0, 1)
+ablines!(ax, 0, 1)
 f
 ```
 """