+++ src/lwplsr_avg.jl
@@ -74,7 +74,7 @@ fm = lwplsr_avg(Xtrain, ytrain; nlvdis = nlvdis,
 res = predict(fm, Xtest)
 rmsep(res.pred, ytest)
 f, ax = scatter(vec(res.pred), ytest)
-abline!(ax, 0, 1)
+ablines!(ax, 0, 1)
 f
 
 fm = lwplsr_avg(Xtrain, ytrain; nlvdis = nlvdis,