+++ src/plskern.jl
@@ -97,7 +97,7 @@ res = predict(fm, Xtest)
 res.pred
 rmsep(res.pred, ytest)
 f, ax = scatter(vec(res.pred), ytest)
-abline!(ax, 0, 1)
+ablines!(ax, 0, 1)
 f
 
 res = predict(fm, Xtest; nlv = 1:2)