+++ src/plsr_avg.jl
@@ -83,7 +83,7 @@ res = predict(fm, Xtest)
 res.pred
 rmsep(res.pred, ytest)
 f, ax = scatter(vec(res.pred), ytest)
-abline!(ax, 0, 1)
+ablines!(ax, 0, 1)
 f
 
 fm = plsr_avg(Xtrain, ytrain; nlv = nlv,