+++ src/rr.jl
@@ -65,7 +65,7 @@ res = predict(fm, Xtest)
 res.pred
 rmsep(res.pred, ytest)
 f, ax = scatter(vec(res.pred), ytest)
-abline!(ax, 0, 1)
+ablines!(ax, 0, 1)
 f
 
 # Only for rr